Scripts/requirements.txt
rich
inquirer
PyYAML